requirements.txt
yt-dlp>=2023.12.30
openai-whisper>=20231117
requests>=2.31.0
